// src/generated/model/escaneoSalida.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Automatizacion de Correos API
 * Ingesta de adjuntos de correo y extraccion de datos tributarios. Todas las rutas requieren un token OIDC salvo las sondas de salud.
 * OpenAPI spec version: 0.1.0
 */
import type { ContadoresSalida } from './contadoresSalida';

export interface EscaneoSalida {
  codigo_de_error: string | null;
  contadores: ContadoresSalida;
  encolado_en: string;
  estado: string;
  fase: string;
  finalizado_en: string | null;
  id: string;
  iniciado_en: string | null;
  mensaje_de_error: string | null;
  progreso_porcentaje: number;
}